requirements.txt
pycryptodomex == 3.8.1